refactor(types): decompose yaml_store.rs into focused modules [OPE-383]
Split the monolithic 510-line yaml_store.rs into a focused module
structure under yaml_store/:

- mod.rs: YamlDefinition trait, YamlFileStore struct with CRUD
  operations and mtime-based caching
- validation.rs: path safety helpers (validated_dir,
  validate_store_path, store_path_for)
- tests.rs: all 21 test cases migrated intact

Public API (YamlDefinition, YamlFileStore) and downstream behavior
preserved. All tests pass, full workspace compiles.
