--- crates/opengoose-types/src/yaml_store.rs
@@ -1,510 +1,0 @@
-use std::collections::HashMap;
-use std::path::{Component, Path, PathBuf};
-use std::sync::{Arc, RwLock};
-use std::time::SystemTime;
-
-/// Trait for YAML-serializable definitions (profiles, teams, etc.)
-///
-/// Implementors provide YAML parsing/serialization and validation logic.
-/// The `YamlFileStore` uses this trait for generic CRUD operations.
-pub trait YamlDefinition: Sized {
-    /// The error type for parsing/validation/serialization.
-    ///
-    /// Must convert from `std::io::Error` so file operations can use `?`.
-    type Error: From<std::io::Error>;
-
-    /// The display name / title of this definition.
-    fn title(&self) -> &str;
-
-    /// Parse from a YAML string, including validation.
-    fn from_yaml(yaml: &str) -> Result<Self, Self::Error>;
-
-    /// Serialize to a YAML string.
-    fn to_yaml(&self) -> Result<String, Self::Error>;
-
-    /// File-safe name derived from the title.
-    fn file_name(&self) -> String {
-        format!(
-            "{}.yaml",
-            crate::sanitize_name(&self.title().to_lowercase().replace(' ', "-"))
-        )
-    }
-}
-
-/// Generic YAML-file-backed CRUD store.
-///
-/// Provides common file operations for any `YamlDefinition` type.
-/// Crate-specific stores (ProfileStore, TeamStore) wrap this with
-/// their own error handling and convenience methods.
-///
-/// Each individual file read is cached in memory and invalidated
-/// automatically when the file's last-modified timestamp changes,
-/// providing hot-reload behaviour with zero extra I/O on cache hits.
-///
-/// `Clone` is cheap: clones share the same underlying `file_cache` `Arc`,
-/// so all copies benefit from each other's reads automatically.
-#[derive(Clone)]
-pub struct YamlFileStore {
-    dir: PathBuf,
-    /// Per-file content cache: path → (raw YAML string, last-modified time).
-    file_cache: Arc<RwLock<HashMap<PathBuf, (String, SystemTime)>>>,
-}
-
-impl YamlFileStore {
-    /// Create a store backed by the given directory.
-    pub fn new(dir: PathBuf) -> Self {
-        Self {
-            dir,
-            file_cache: Arc::new(RwLock::new(HashMap::new())),
-        }
-    }
-
-    /// Read a file through the in-memory cache.
-    ///
-    /// Returns cached content when the file's mtime matches; re-reads
-    /// from disk and updates the cache otherwise.
-    fn read_cached(&self, path: &Path) -> std::io::Result<String> {
-        self.validate_store_path(path)?;
-        let mtime = std::fs::metadata(path)?.modified()?;
-
-        // Fast path: valid cache entry.
-        if let Ok(cache) = self.file_cache.read()
-            && let Some((content, cached_mtime)) = cache.get(path)
-            && *cached_mtime == mtime
-        {
-            return Ok(content.clone());
-        }
-
-        // Cache miss or stale — read from disk and update.
-        let content = std::fs::read_to_string(path)?;
-        if let Ok(mut cache) = self.file_cache.write() {
-            cache.insert(path.to_path_buf(), (content.clone(), mtime));
-        }
-        Ok(content)
-    }
-
-    /// Evict a single file from the cache (called after writes/deletes).
-    fn invalidate(&self, path: &Path) {
-        if let Ok(mut cache) = self.file_cache.write() {
-            cache.remove(path);
-        }
-    }
-
-    /// The directory path this store operates on.
-    pub fn dir(&self) -> &Path {
-        &self.dir
-    }
-
-    /// Ensure the backing directory exists.
-    pub fn ensure_dir(&self) -> std::io::Result<()> {
-        std::fs::create_dir_all(self.validated_dir()?)
-    }
-
-    /// List all definition names (sorted) by scanning YAML files.
-    pub fn list<T: YamlDefinition>(&self) -> Result<Vec<String>, T::Error> {
-        let dir = self.validated_dir()?;
-        if !dir.exists() {
-            return Ok(vec![]);
-        }
-
-        let mut names = Vec::new();
-        for entry in std::fs::read_dir(dir)? {
-            let entry = entry?;
-            let path = entry.path();
-            self.validate_store_path(&path)?;
-            if path
-                .extension()
-                .is_some_and(|ext| ext == "yaml" || ext == "yml")
-                && let Ok(content) = std::fs::read_to_string(&path)
-                && let Ok(item) = T::from_yaml(&content)
-            {
-                names.push(item.title().to_string());
-            }
-        }
-        names.sort();
-        Ok(names)
-    }
-
-    /// Get a definition by name.
-    ///
-    /// Reads through the in-memory file cache; disk access only occurs
-    /// when the file has changed since the last read (mtime-based hot reload).
-    pub fn get<T: YamlDefinition>(&self, name: &str) -> Result<T, T::Error> {
-        let path = self.store_path_for(name)?;
-        let content = self.read_cached(&path)?;
-        T::from_yaml(&content)
-    }
-
-    /// Save a definition. If `force` is false and the file exists,
-    /// returns `Err` with an `io::ErrorKind::AlreadyExists` error.
-    pub fn save<T: YamlDefinition>(&self, item: &T, force: bool) -> Result<(), T::Error> {
-        self.ensure_dir()?;
-        let path = self.store_path_for(item.title())?;
-        if !force && path.exists() {
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::AlreadyExists,
-                format!("'{}' already exists", item.title()),
-            )
-            .into());
-        }
-        let yaml = item.to_yaml()?;
-        std::fs::write(&path, yaml)?;
-        self.invalidate(&path);
-        Ok(())
-    }
-
-    /// Remove a definition by name.
-    pub fn remove(&self, name: &str) -> std::io::Result<()> {
-        let path = self.store_path_for(name)?;
-        if !path.exists() {
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::NotFound,
-                format!("'{name}' not found"),
-            ));
-        }
-        std::fs::remove_file(&path)?;
-        self.invalidate(&path);
-        Ok(())
-    }
-
-    /// Resolve the file path for a definition name.
-    ///
-    /// Uses `sanitize_name` to prevent path traversal.
-    pub fn path_for(&self, name: &str) -> PathBuf {
-        let sanitized = crate::sanitize_name(&name.to_lowercase().replace(' ', "-"));
-        self.dir.join(format!("{sanitized}.yaml"))
-    }
-
-    fn store_path_for(&self, name: &str) -> std::io::Result<PathBuf> {
-        let dir = self.validated_dir()?;
-        let sanitized = crate::sanitize_name(&name.to_lowercase().replace(' ', "-"));
-        let path = dir.join(format!("{sanitized}.yaml"));
-        self.validate_store_path(&path)?;
-        Ok(path)
-    }
-
-    fn validated_dir(&self) -> std::io::Result<PathBuf> {
-        if self.dir.as_os_str().is_empty() {
-            return Err(invalid_store_path("store directory must not be empty"));
-        }
-
-        let mut validated = PathBuf::new();
-        for component in self.dir.components() {
-            match component {
-                Component::Prefix(prefix) => validated.push(prefix.as_os_str()),
-                Component::RootDir => validated.push(Path::new(std::path::MAIN_SEPARATOR_STR)),
-                Component::Normal(part) => validated.push(part),
-                Component::CurDir | Component::ParentDir => {
-                    return Err(invalid_store_path(
-                        "store directory must not contain relative traversal components",
-                    ));
-                }
-            }
-        }
-
-        Ok(validated)
-    }
-
-    fn validate_store_path(&self, path: &Path) -> std::io::Result<()> {
-        let dir = self.validated_dir()?;
-        let relative = path.strip_prefix(&dir).map_err(|_| {
-            invalid_store_path("store paths must stay within the configured store directory")
-        })?;
-        let mut components = relative.components();
-        match (components.next(), components.next()) {
-            (Some(Component::Normal(_)), None) => Ok(()),
-            _ => Err(invalid_store_path(
-                "store paths must resolve to a single file inside the store directory",
-            )),
-        }
-    }
-}
-
-fn invalid_store_path(message: &str) -> std::io::Error {
-    std::io::Error::new(std::io::ErrorKind::InvalidInput, message)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Simple test implementation of YamlDefinition.
-    #[derive(Debug, Clone, PartialEq)]
-    struct TestDef {
-        name: String,
-        value: String,
-    }
-
-    impl YamlDefinition for TestDef {
-        type Error = std::io::Error;
-
-        fn title(&self) -> &str {
-            &self.name
-        }
-
-        fn from_yaml(yaml: &str) -> Result<Self, Self::Error> {
-            let mut name = String::new();
-            let mut value = String::new();
-            for line in yaml.lines() {
-                if let Some(rest) = line.strip_prefix("name: ") {
-                    name = rest.to_string();
-                } else if let Some(rest) = line.strip_prefix("value: ") {
-                    value = rest.to_string();
-                }
-            }
-            if name.is_empty() {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::InvalidData,
-                    "missing name",
-                ));
-            }
-            Ok(TestDef { name, value })
-        }
-
-        fn to_yaml(&self) -> Result<String, Self::Error> {
-            Ok(format!("name: {}\nvalue: {}\n", self.name, self.value))
-        }
-    }
-
-    fn test_def(name: &str, value: &str) -> TestDef {
-        TestDef {
-            name: name.to_string(),
-            value: value.to_string(),
-        }
-    }
-
-    #[test]
-    fn test_new_and_dir() {
-        let dir = PathBuf::from("/tmp/test-store");
-        let store = YamlFileStore::new(dir.clone());
-        assert_eq!(store.dir(), dir.as_path());
-    }
-
-    #[test]
-    fn test_ensure_dir() {
-        let tmp = tempfile::tempdir().unwrap();
-        let nested = tmp.path().join("a").join("b");
-        let store = YamlFileStore::new(nested.clone());
-        assert!(!nested.exists());
-        store.ensure_dir().unwrap();
-        assert!(nested.exists());
-    }
-
-    #[test]
-    fn test_path_for() {
-        let store = YamlFileStore::new(PathBuf::from("/store"));
-        assert_eq!(
-            store.path_for("My Profile"),
-            PathBuf::from("/store/my-profile.yaml")
-        );
-    }
-
-    #[test]
-    fn test_path_for_traversal() {
-        let store = YamlFileStore::new(PathBuf::from("/store"));
-        let path = store.path_for("../../etc/passwd");
-        assert!(path.starts_with("/store"));
-        assert!(!path.to_string_lossy().contains(".."));
-    }
-
-    #[test]
-    fn test_file_name_trait_default() {
-        let def = test_def("My Cool Profile", "v1");
-        assert_eq!(def.file_name(), "my-cool-profile.yaml");
-    }
-
-    #[test]
-    fn test_save_and_get() {
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-        let item = test_def("alpha", "one");
-        store.save(&item, false).unwrap();
-        let loaded: TestDef = store.get("alpha").unwrap();
-        assert_eq!(loaded, item);
-    }
-
-    #[test]
-    fn test_list_empty() {
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-        let names = store.list::<TestDef>().unwrap();
-        assert!(names.is_empty());
-    }
-
-    #[test]
-    fn test_list_nonexistent_dir() {
-        let store = YamlFileStore::new(PathBuf::from("/nonexistent/path/xyz"));
-        let names = store.list::<TestDef>().unwrap();
-        assert!(names.is_empty());
-    }
-
-    #[test]
-    fn test_list_with_items() {
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-        store.save(&test_def("charlie", "3"), false).unwrap();
-        store.save(&test_def("alpha", "1"), false).unwrap();
-        store.save(&test_def("bravo", "2"), false).unwrap();
-        let names = store.list::<TestDef>().unwrap();
-        assert_eq!(names, vec!["alpha", "bravo", "charlie"]);
-    }
-
-    #[test]
-    fn test_save_no_force_duplicate() {
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-        store.save(&test_def("item", "v1"), false).unwrap();
-        let err = store.save(&test_def("item", "v2"), false).unwrap_err();
-        assert_eq!(err.kind(), std::io::ErrorKind::AlreadyExists);
-    }
-
-    #[test]
-    fn test_save_force_overwrites() {
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-        store.save(&test_def("item", "v1"), false).unwrap();
-        store.save(&test_def("item", "v2"), true).unwrap();
-        let loaded: TestDef = store.get("item").unwrap();
-        assert_eq!(loaded.value, "v2");
-    }
-
-    #[test]
-    fn test_remove() {
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-        store.save(&test_def("item", "v1"), false).unwrap();
-        store.remove("item").unwrap();
-        assert!(store.get::<TestDef>("item").is_err());
-    }
-
-    #[test]
-    fn test_remove_not_found() {
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-        let err = store.remove("nonexistent").unwrap_err();
-        assert_eq!(err.kind(), std::io::ErrorKind::NotFound);
-    }
-
-    #[test]
-    fn test_get_not_found() {
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-        let err = store.get::<TestDef>("nonexistent").unwrap_err();
-        assert_eq!(err.kind(), std::io::ErrorKind::NotFound);
-    }
-
-    #[test]
-    fn test_list_ignores_non_yaml_files() {
-        // list() only considers files with .yaml or .yml extension.
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-
-        // Write a valid definition through the store (gets a .yaml extension).
-        store.save(&test_def("alpha", "1"), false).unwrap();
-
-        // Write files with other extensions directly — these must be skipped.
-        std::fs::write(
-            tmp.path().join("README.txt"),
-            "name: should-be-ignored\nvalue: x\n",
-        )
-        .unwrap();
-        std::fs::write(
-            tmp.path().join("config.json"),
-            "name: also-ignored\nvalue: y\n",
-        )
-        .unwrap();
-        std::fs::write(tmp.path().join("no-extension"), "name: no-ext\nvalue: z\n").unwrap();
-
-        let names = store.list::<TestDef>().unwrap();
-        assert_eq!(names, vec!["alpha"]);
-    }
-
-    #[test]
-    fn test_list_ignores_invalid_yaml_files() {
-        // list() silently skips .yaml files that fail to parse as valid definitions.
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-
-        // Write a valid item.
-        store.save(&test_def("valid", "ok"), false).unwrap();
-
-        // Write a .yaml file that will fail TestDef::from_yaml (missing "name" field).
-        std::fs::write(
-            tmp.path().join("broken.yaml"),
-            "value: only-value-no-name\n",
-        )
-        .unwrap();
-
-        let names = store.list::<TestDef>().unwrap();
-        assert_eq!(names, vec!["valid"]);
-    }
-
-    #[test]
-    fn test_read_cached_returns_cached_on_same_mtime() {
-        // Exercise the cache-hit path (lines 72-74): a second `get()` for the
-        // same file, without any modification in between, should return the
-        // cached content without re-reading from disk.
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-        let item = test_def("cached-item", "v1");
-        store.save(&item, false).unwrap();
-
-        // First read → cache miss, populates cache.
-        let first: TestDef = store.get("cached-item").unwrap();
-        assert_eq!(first.value, "v1");
-
-        // Second read → cache hit (mtime unchanged).
-        let second: TestDef = store.get("cached-item").unwrap();
-        assert_eq!(second.value, "v1");
-    }
-
-    #[test]
-    fn test_cache_invalidated_after_save() {
-        // After `save(force=true)`, the cache entry is evicted, so the next
-        // `get()` must re-read from disk and see the updated content.
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-
-        store.save(&test_def("item", "v1"), false).unwrap();
-        let first: TestDef = store.get("item").unwrap();
-        assert_eq!(first.value, "v1");
-
-        // Overwrite via force-save.
-        store.save(&test_def("item", "v2"), true).unwrap();
-        let second: TestDef = store.get("item").unwrap();
-        assert_eq!(second.value, "v2");
-    }
-
-    #[test]
-    fn test_list_includes_yml_extension() {
-        // list() accepts both .yaml and .yml extensions.
-        let tmp = tempfile::tempdir().unwrap();
-        let store = YamlFileStore::new(tmp.path().to_path_buf());
-
-        store.save(&test_def("via-store", "1"), false).unwrap();
-        // Write a valid item with .yml extension directly.
-        std::fs::write(
-            tmp.path().join("shortform.yml"),
-            "name: shortform\nvalue: 2\n",
-        )
-        .unwrap();
-
-        let mut names = store.list::<TestDef>().unwrap();
-        names.sort();
-        assert_eq!(names, vec!["shortform", "via-store"]);
-    }
-
-    #[test]
-    fn test_ensure_dir_rejects_parent_dir_components() {
-        let store = YamlFileStore::new(PathBuf::from("../unsafe-store"));
-        let err = store.ensure_dir().unwrap_err();
-        assert_eq!(err.kind(), std::io::ErrorKind::InvalidInput);
-    }
-
-    #[test]
-    fn test_get_rejects_nested_store_path() {
-        let store = YamlFileStore::new(PathBuf::from("/tmp/store/nested"));
-        let err = store.validate_store_path(Path::new("/tmp/store/nested/inner/item.yaml"));
-        assert!(err.is_err());
-    }
-}
